Update patch set 8
Patch Set 8:

> Just gave a quick look, but I like introduction of
 > PushMode.getWizard().

If you can spare the time I'd appreciate your feedback on the choice of Gerrit wizard/Push branch wizard. For a Gerrit repo, it's still tied to the state of the "Change-ID" button. Which is of course an approximation only. First, Gerrit might be configured to allow pushing patch sets without change id, and second, someone might want to push a commit with change id directly.

We could also add a link in the Gerrit push wizard to open the Push branch wizard (like in the push branch wizard to get to the full general push wizard).

Patch-set: 8
Label: Verified=0