Update patch set 8
Patch Set 8:

> If you can spare the time I'd appreciate your feedback on the choice of Gerrit wizard/Push branch wizard. For a Gerrit repo, it's still tied to the state of the "Change-ID" button. Which is of course an approximation only.

As mentioned earlier, I'd personally be fine with no Gerrit wizard at all ;) But the approximation you implemented seems sane and is IMO good enough for a merge soon. IMO, it's worth pushing the change as it, let people play with it from snapshots and see whether it seems good enough for most people or whether it's something we should change in some future commit.

> We could also add a link in the Gerrit push wizard to open the Push branch wizard (like in the push branch wizard to get to the full general push wizard).

That would be nice. Would you like me to open a Bugzilla issue about it?

Patch-set: 8