Block commit on error
The previous change removed the initial condition "commitEnabled" when
inlining the method for calculating the button enablements. That broke
the "block commit" functionality.

Re-add the flag, but ignore it in case of "push head only" mode. That
way staging a file with error leads to both buttons blocked, unstaging
that file leads to an enabled "Push HEAD" button.

Change-Id: I4899e27ad158ac746c41b778861c71c3e42473a4
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>
Bug: 563512