WIP: StagingView: use virtual Tree to improve UI performance
Thousands (but less than 10k) of files to show in the staging view
consumes lots of time in the UI thread, because setting Text and Image
for the TreeItem per StagingEntry fires JNI calls to the underlying SWT
framework.
Instead of initializing all TreeItems upfront, they can be initialized
when becoming visible. For the TreeViewer, this is the SWT.VIRTUAL style
and the SetData listener on the underlying Tree control, not using the
TreeViewer.setInput() API, because TreeViewer's support for lazy content
providers does not support sorting and filtering.

WIP: handle Selections, switch Presentation mode, Sorting, Filtering

Change-Id: Ifd8cba096a315d3931ac907b3607fff49a36fef4
Signed-off-by: Michael Haubenwallner <michael.haubenwallner@ssi-schaefer.com>