[staging] Clear and restore selection when refreshing
Clear the selection before refreshing a view, and re-set it
afterwards. refreshViewersInternal() is always called inside a
setRedraw(false), so this has no visible effects. But it greatly
simplifies the work of viewer.refresh(), which might otherwise try
to re-set the selection too many times, which causes quadratic
behavior when the presentation is switched from tree view to list
view while there is a selection.

Bug: 564569
Change-Id: Ib684b15642aeb8bc5ec8a357db7a8546732a48cd
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>