Bug 564708 - Migrate Jobs to DS
The template resolver and the configuration checker in the Activator
needs to be started when the EGit bundle is started, but does not need
to be done in the Activator itself.  By externalising this as a DS
component, it will be started when the bundle starts but asynchronously
from the Activator content, thus will reduce the amount of time required
to start EGit.

Moving these to Jobs and enabling them as components allows them to be
independently controlled, and terminated when the bundle or component is
stopped.

Change-Id: I7acb05c3de6baf020735dc85ef589f6ac69c632d
Signed-off-by: Alex Blewitt <alex.blewitt@gmail.com>