[dark mode] Fix background color of unstaged/ignored resources
Change the color to #2F2F2F (decimal 47; was #2A2A2A, decimal 42).
Even on Neon, dark mode uses #2F2F2F (on Mac) for the background of
trees. #2A2A2A used to be the background of editors/StyledText for
some time.

The history of the Eclipse dark mode CSS files is convoluted and not
quite clear to me with all the simplifications that have occurred and
different CSS for different platforms, but I didn't see any version
that used #2A as background for trees.

So just change the value in our dark mode CSS and hope it'll be all
right not just on Eclipse Neon and 4.17 but also on all version in
between.

Bug: 564478
Change-Id: Iecfcf21353979e9938cf05882accfe5ca8933cf6
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>