Update patch set 1
Patch Set 1: Code-Review+1

This looks good to me both in terms of implementation & user testing.
I tried this change with a build of org.eclipse.ui that was missing the TREE_TABLE_FONT and a build that included it, and didn't see any exceptions or issues in either case. 

One thing that I just noticed: the list of changes to stage/unstage in the git staging view seems affected by the TREE_TABLE_FONT. I would assume this is desirable but wanted to point it out in any case.

Patch-set: 1
Reviewer: Gerrit User 246338 <246338@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=+1