[staging view] Use virtual Tree to improve UI performance
Thousands (but less than 10k) of files to show in the staging view
consumes lots of time in the UI thread, because setting Text and Image
for the TreeItem per StagingEntry fires JNI calls to the underlying SWT
framework, and querying the item decorations is expensive by itself.
Instead of initializing all TreeItems up front, they can be initialized
when becoming visible.
For the TreeViewer, this is the SWT.VIRTUAL style and the SetData
listener on the underlying Tree control, not using the
TreeViewer.setInput() API.

This commit does introduce the VIRTUAL TreeViewer and the SetData
listener, not supporting features like switching presentation mode,
handle selections, sorting, filtering and similar yet.

Change-Id: Ifd8cba096a315d3931ac907b3607fff49a36fef4