[staging view] Use asynchronous lightweight decorations
Simplify the StagingViewLabelProvider not to decorate and not use
StyledString in getText(), which returns a plain String anyway.

Change the label provider of the staged/unstaged viewers to use the
platform's asynchronous lightweight decorator mechanism, which will
use EGit's GitLightweightDecorator to decorate the entries. We still
have to add the marker decorations explicitly. That is to be
investigated in the future.

This gets rid of a problematic direct use of DecorationHelper outside
of the GitLightweightDecorator in preparation of a solution for bug
565851.

Change-Id: Idd49496ad42ee9e6d47524adb968091c5b65b25b
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>