[history] Less flashy "other references" in dark mode
"Other references" like HEAD, FETCH_HEAD, and so on were always
shown black on near-white. That doesn't look well in dark mode.

Newly the rendering adapts to the table background, which changes
when the theme changes. If the table has a dark background use
white text on a dark background (slightly lighter than the table
background). Otherwise use black text on white background.

Bug: 536171
Change-Id: I0fdb9b5d6ecf0e16b221f559c33c7d6a649b401a
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>