Update patch set 1
Patch Set 1:

The build failed because of a bug in the build scripts. The build picked JGit 5.1.3 as upstream instead of 5.10.0.

See the fix in https://git.eclipse.org/r/c/egit/egit-pipelines/+/168404 . We don't use egit-pipelines yet, but I've made the same fix manually in all EGit jobs, and have retriggered this build, which then picked the correct JGit 5.10.0 to build against.

Patch-set: 1
CC: Gerrit User 44628 <44628@97ee7c02-f12f-4043-b43e-dea463d88b31>