[staging view] Reduce redraws
Updating the warning label may require a redraw of the whole commit
message section because otherwise the content assist lightbulb may not
be redrawn correctly if the label appears or disappears. However, the
staging view did redraw way too often -- if neither the visibility nor
the message nor the image change, a redraw is not needed.

Improve the ToggleableLabel to make all methods indicate via their
return value whether something changed and a redraw/relayout may be
needed. Then the user of the label can decide whether to redraw or to
re-layout.

Previously, the staging view would redraw on each keypress in the
commit message text area; now it only redraws when the warning message
changes.

Bug: 565019
Change-Id: Ifc3481bec77023ec3c722922694aa9955576806d
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>