Update patch set 2
Patch Set 2:

If the new preference is switched off and I clone a repository via ssh the passphrase dialog comes up asking for the passphrase. When I enter it and check the checkbox to store the passphrase in the secure store I'd expect that this switches on the new preference and that I am not asked again for the same passphrase again within the same clone wizard.

Patch-set: 2
CC: Gerrit User 4 <4@97ee7c02-f12f-4043-b43e-dea463d88b31>