[gerrit fetch] Move clipboard handling to command
FetchGerritChangePage takes a constructor parameter for the initial
value, but it was always null. The FetchGerritChangeWizard constructor
passing it on was never used.

Move the clipboard handling to the command and get the clipboard text
before we possibly open a repository selection dialog. Then pass on
that clipboard text to the wizard and thus the page. The main benefit
of this change is a minor simplifcation of the very complex
FetchGerritChangePage.

Change-Id: I87e44b9a61fa27f0b3e9fe934c6fe29c979bb2bb
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>