Enable "Replace with HEAD" in more cases
The action was enabled only for RepositoryState.SAFE, meaning one
could not use it during a rebase or merge. But actually that condition
is unnecessarily restrictive; the operation can be allowed anytime
provided the repository is not bare and has a HEAD.

Improve the property testers to be able to express the condition "all
selected items are in a repository, which is non-bare and has a head"
efficiently: allow tester arguments to be prefixed by "!" to signify
logical negation. Add a new property test for "hasHead". Then change
the repository test in the enablement condition of the command from
"isSafe" to "!isBare,hasHead". (Multiple arguments are ANDed.)

Adapt the handler's isEnabled(), and add a UI test using "Replace with
HEAD" to resolve a conflict.

Bug: 568038
Change-Id: I7e586c3fa35d0c5b66c8d07a9504105d0765ce77
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>