Update patch set 1
Patch Set 1:

this behaves different than checkout HEAD -- <path> in C git when applied on a conflicting file:

EGit/JGit not only replace the content of the selected file with the content from HEAD
but also stages this modification. C git only replaces the content but still requires
an explicit git add to stage this modification and mark the conflict resolved.

Patch-set: 1
CC: Gerrit User 4 <4@97ee7c02-f12f-4043-b43e-dea463d88b31>