Update patch set 1
Patch Set 1:

I don't quite understand. mylyn.commons was never in Orbit AFAIK? It was in simrel... At bug 564699 you wrote "otherwise we have to remove it from simrel": "it" being the egit mylyn integration? But we'd leave them in our p2 repo? What happens if a user upgrades from egit 5.9? I would suspect that p2 doesn't find a solution :-(.

Are these problems with xml.bind also the reason for not updating the Orbit link in the EGit p2 repo?

Patch-set: 1
CC: Gerrit User 44628 <44628@97ee7c02-f12f-4043-b43e-dea463d88b31>