Update patch set 1
Patch Set 1:

> I thought this might be useful for installing from a download of the p2 repos zip file on machines which don't have network access.

Agreed; I forgot about that case.

> Not sure if the dependencies should be in a category or not. WDYT ?

In Eclipse the dialog for installing will by default not show anything that is not in a category. Of course a user can toggle this and see the bundles anyway. Don't know if not having a category might have side-effects. It was just an idea; it's OK for me if these bundles do have a category.

Patch-set: 1