Reword a commit without the rebase interactive machinery
Using interactive rebase for rewording a commit has the disadvantage
that it first rewinds the working tree and then rebases the whole
history from the commit amended to HEAD. This leads to resource change
events and triggers potentially time-consuming builds.

This is not necessary. Rewording a commit message does not change
anything in the working tree or in the index, and it can be done as a
pure repository operation without touching either.

Re-implement RewordCommitOperation by doing the whole rebase
non-interactively in memory: simply create a copy of the original
commit with the new message, then create copies of the children,
patching in new created commits in place of former parents. At the
end, update HEAD and ORIG_HEAD.

While doing so, re-sign commits that were signed and committed by the
same committer.

As a side effect, this new way can also reword merge commit messages.

Bug: 562602
Change-Id: I50dbfa2f4f8246b0c59ef03ad8dd7a3e465fa9d9
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>