Update patch set 2
Patch Set 2:

I did this in EGit instead of in JGit on purpose. I'm not sure this mechanism is general enough to have in JGit. For instance, it only rebases the history between HEAD and the commit being reworded. If there's a merge from this branch onto some other branch, that other branch is not rewritten. Maybe that'd be fine in general, or maybe not. Or maybe a general operation should allow modifying any commit anywhere... Another consideration for a more general operation is whether tags pointing at commits that get rewritten should also be rewritten. Also, the handling of signatures would need more thought for a general operation. (For EGit, we could give the user more control by adding things to the CommitMessageEditorDialog, compare bug 440211.) But all this goes way beyond the scope of this change and beyond what EGit's RewordCommitOperation did before.

I notice that EGit's way of restricting this to commits reachable from HEAD may traverse the whole history if the commit is _not_ reachable from HEAD. Not sure how to get around this; maybe additionally use the commit time as a cut-off? (If the commit to be changed was committed at T0, we might stop the traversal at commits older than T0, perhaps plus some saftey margin?) And unless I misunderstand something, CommitUtil.isCommitInCurrentBranch() will internally traverse the whole history anyway. I don't see anything telling the walk to stop after it has found the commit.

This history rewriting is of course not limited to changing the message of a commit. It could just as well be used for any other modification that doesn't change the trees (change author/committer/dates/signatures).

Patch-set: 2
Reviewer: Gerrit User 4 <4@97ee7c02-f12f-4043-b43e-dea463d88b31>