Support signed tags
JGit's TagCommand has learned to sign tags. Add a checkbox to the
CreateTagDialog where the user can specify whether the new tag shall
be signed. Information for signing is taken from the git config. The
checkbox is added only if signing is possible at all -- if we can't
find a signing key, tags will never be signed, and the checkbox is
not shown at all.

A signed tag is always an annotated tag, even if the message is empty.

Rewrite TagOperation to make use of JGit's TagCommand, and hide the
complexities of setting the correct parameters for TagCommand. Adapt
all users of TagOperation to use the new interface.

Change-Id: I78162dc5a1c51cc345eaab9d801e21011a1a3f52
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>