CreateTagDialog: fix determining an existing tag
setExistingTag() expects a RevTag (annotated tag) or a Ref for a
lightweight tag. However, it was given a RevCommit in the latter
case, and then disabled the "force replace existing tag" checkbox
because it handled that as a non-existing tag.

Be more careful when resolving the tag ref, and pass the Ref when it
resolves to a non-RevTag object (in practice, to a RevCommit).

Bug: 569653
Change-Id: If1588b28475afdf540d13eaaab93cbafcdc007a9
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>