[tag properties] Show full message in a pop-up
Show the full message of a tag or commit in the properties view in
an IInformationControl when the property is selected, similar to the
information hover pop-ups in text editors.

The IInformationControl framework from JFace needs quite a bit of setup,
though. Although the JFace support is prepared for being used with
arbitrary controls, the Eclipse platform provides only built-in support
for text editors.

Change-Id: Ieef3b22044b956e5368f962206a68a4df86f6853
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>