Link the commit viewer with the properties view
Move all the new property sheet related classes that work with general
git objects into a new package, and enable support for the property
view also in the commit viewer.

When the commit viewer is active, the property view shows the data of
that commit. With the properties view visible the user can thus see
information about the commit even when the diff or notes page is
active.

Change-Id: I62c1c4f19b31f78aad3dce63d552796f0ebd51ea
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>