Update patch set 3
Patch Set 3:

Looking at the debug output, more precise steps seem to be:

    1. Open Package Explorer, Git Repositories view, git History view page, Error Log view in separate view stacks.
    2. Select a tag from the Git Repositories view to populate the git History page.
    3. Select a commit in the commit graph, select text in the commit viewer.
    4. Click on the Package Explorer area.
    5. Click on another tag in the Git Repositories view.
    6. Select text in the commit viewer.
    7. Select a commit in the commit graph.

Repeat once if the problem is not seen. I'm not 100% those are the steps, due to having to repeat them. But this reproduces the problem more or less fast.

What I see is (I think), the FocusTracker will store a StyledText widget as the last widget to have focus, upon clicking on the Package Explorer view area. When clicking on a tag in the Git Repositories, an activation event is sent to the git History view page, on some of the widgets that have actions attached to them (this should be coming from my merged change for bug 536645). Then focus is sent to the git History view page, due to clicking on a commit; this results in the FocusTracker judging the focus must go to the StyledText widget (the commit viewer). Then the 2nd handler is activated and the conflict occurs.

Patch-set: 3
Reviewer: Gerrit User 119988 <119988@97ee7c02-f12f-4043-b43e-dea463d88b31>