GPG signing: use an external GPG executable
Don't rely on JGit's GPG signer based on BouncyCastle. BouncyCastle
in Eclipse is crippled; it lacks the patent-encumbered AES/OCB cipher
that is needed to decrypt encrypted private keys stored by GPG in the
Extended Key Format. Also, the JGit BouncyCastle GpgSigner has to rely
on a number of assumptions about GPG internals. Compare bug 570501.

Instead use a GpgSigner implementation that uses an external GPG
executable for signing. That way it is guaranteed that signing works
if it works in command-line git. Delegate all passphrase handling to
the external GPG. This gives automatic integration with the native
keychain and with gpg-agent.

Add a new EGit preference at Git->Committing where the user can specify
the path to a GPG executable. EGit finds the GPG to use

* via the EGit preference, if set, otherwise
* via the git config gpg.program, if set, otherwise
* by looking on $PATH for an executable named "gpg" (or "gpg.exe").

If the EGit preference is an invalid path or not an executable file,
errors are logged and the EGit preference is ignored. Normally, it
should not be necessary to set the Eclipse preference, but it helps
avoid the lookup on $PATH, and it may help if for some reason the
value specified in the git config should not work with EGit.

Verifying signatures still uses the JGit verifier, which is also based
on BouncyCastle. Verifying doesn't need AES/OCB, and needs only a
public key, which JGit can find without access to GPG internals.

Automated tests for this are not possible on the current build
infrastructure as it would need a GPG installation and keys. The
change has been tested manually on

* OS X 10.14.6 with gpg 2.2.25
* CentOS 7 with gpg 2.0.22
* Win 10 with gpg 2.2.27 (gpg4win 3.1.15)

Bug: 547789
JGit-Dependency: If7e34aeed6ca6636a92bf774d893d98f6d459181
Change-Id: Id95b89cfbf822422668f668a316c5a2d8ee2d847
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>