Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> I'd like to see the basic of GPG tool based signing to be part of JGit. EGit would just configure it by default.
> 
> Thus, should ExternGpgSigner be part of JGit?

See bug 570501. I think that JGit should remain as much Java-only as possible. Calling gpg the way it's done here as

  gpg -bsau <keySpec> --batch --no-tty --status-fd 2 --output -

is fine for a graphical program like Eclipse.

If you use JGit as a library or jgit pgm in a script, you'd probably want to be able to configure a CredentialsProvider to avoid any GUI prompt, which requires calling gpg differently. Basically you'd have to write the data to be signed to a temporary file and then use

  gpg -bsau <keySpec> --batch --no-tty --pinentry-mode loopback --password-fd 0 --status-fd 2 --output - <tmp_file_name>

and write the password through stdin (in the platform default encoding?). --pinentry-mode is available only as of GPG 2.1. And that way has the problem that you the CredentialsProvider would have to ask for a passphrase up front, even if the key used in the end is unprotected. I haven't found a way to make gpg tell the Java code when it wants what kind of password.

Patch-set: 3