Update patch set 1
Patch Set 1:

Updated all EGit builds to use v0.0.6 (which is the isUnix() fix). The EGit product builds are fine checking GERRIT_BRANCH; they trigger on "change merged", and for that GERRIT_BRANCH is set. See e.g. [1]. The JGit product builds trigger on "ref changed", and for that GERRIT_BRANCH is not set, but GERRIT_REF_NAME is.

[1] https://ci.eclipse.org/egit/job/egit-stable/143/parameters/

Patch-set: 1