Use an OSGi declarative service to manage some EGit components
Add a StartEventHandler to start things that require the workbench to
have been created. If the workbench is already running when the
component is activated, it initializes the EGit components right away.
Otherwise, the handler is notified on the APP_STARTUP_COMPLETED event
and initializes the EGit components then.

This change uses this mechanism for the SelectionRepositoryStateCache
only, but it might be used to manage other EGit internal components in
the future.

Maybe some parts of EGit might themselves become OSGi declarative
services and deal with their own activation as they need. The single
StartEventHandler is a useful bridge facilitating the move to OSGi
declarative services while minimizing code changes, though.

Change-Id: Ic12507ef24dc940a25028ff2e1a1b77f64276078
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>