Update patch set 2
Patch Set 2:

I think the idea of moving the startup code in the Activator into a DS which then listens out for the startup event is a good idea.

The question I have is if any code subsequent to the activator relies on this code having been executed? If so it might need to go in there as well.

I've been looking to see about moving some of the Jobs that get started into DS components as well but I've been delayed in writing it up.

So I think this approach looks good :)

Patch-set: 2
Reviewer: Gerrit User 3 <3@97ee7c02-f12f-4043-b43e-dea463d88b31>