Update patch set 2
Patch Set 2: Code-Review+1

I think this is good to go as is.

I think we can investigate moving other componentns to this at a later stage; for example, the FocusListener depends on the Workbench (and we have a remove in the shutdown of the bundle) which could be handled by this component listening out for the IWorkbench arriving and setting a focus listener, and similarly for removal -- but those could be additional changes after this one.

Patch-set: 2
Reviewer: Gerrit User 3 <3@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=+1