Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> I was also thinking about making the DebugOptionsListener a separate DS, but that would require hard-coding the bundle id somewhere. Probably not worth the trouble. (But Activator.stop() should unregister that DebugOptionsListener? Or does that happen automatically for services registered via the BundleContext?)

The purpose of the unregistration in the activator.stop is to prevent keeping a reference from the listener to the class after the bundle stops; however, when a bundle stops all of its DS components are stopped as well. So there's no need to do it twice; and with luck, we won't have the listener referenced anyway, so we can remove it from the Activator in its entirety.

I had a go at moving some things to Jobs a while ago; I've marked a change as WIP because I think it might be better to use this approach, and merely move the startup into this one component instead of having multiple. But you can see how you could use this to have a service which depends on the resources rather than using static Platform methods:

https://git.eclipse.org/r/c/egit/egit/+/175920
 
> If we end up with multiple services having dependencies amongst each other, I presume injecting one 
service in another tells the framework in which order to activate/deactivate them?

If you have a service A which depends on another B (and it's marked as mandatory), then A won't start until B is present. (You could solve the 'is the workbench available' by depending on the org.eclipse.ui.IWorkbench service as well, but that might be ready before the E4 startup is, so your approach here is better.) If B goes away, then A will automatically shut down.
 
> Is there a way to restrict the visibility of a service? Or can anyone get the service via the service registry?

Anyone can get the service from the registry, but that's sort of the point. The publication of the DebugListener will allow any bundle to publish debug options to EGit -- but this will then allow (for example) a dynamic debug plugin to send in debug opts instead of using a command line interface. The thing is that the EGit just needs to know how to receive debug opts, rather than where they come from.

By the way, if you want to find out the bundle name, you can use FrameworkUtil.getBundle(class) to get the Bundle object, and from there, the getSymbolicName() will give you the answer -- you don't need to hard-code it. See e.g.:

https://github.com/eclipse/eclipse.platform.ui/blob/07db713f17dd4b233f734ef7b902c70908078c67/bundles/org.eclipse.e4.ui.workbench/src/org/eclipse/e4/ui/internal/workbench/DefaultLoggerProvider.java#L34

Patch-set: 3
Reviewer: Gerrit User 3 <3@97ee7c02-f12f-4043-b43e-dea463d88b31>