Update patch set 3
Patch Set 3:

By the way, if you want to look at a DebugOptionsListener you might want to look at this change:

https://github.com/eclipse/epp.mpc/commit/8ebbf4ec0de34b45187e634f136bbfd50818a410

I moved the DebugOptions out from the activator into a (static) inner class so that when it's used for the first time it will assign the entries in there. There was some changes to move from `Activator.debug` to `DebugOptions.debug` but ultimately it allowed to get rid of the Activator (or in this case, PluginUI) in its entirety.

Patch-set: 3
Reviewer: Gerrit User 3 <3@97ee7c02-f12f-4043-b43e-dea463d88b31>