Simplify the EGit UI Activator by using OSGi declarative services
The EGit UI Activator contained a lot of code to deal with external
modifications to git repositories. Refactor this into more manageable
components created dynamically by the framework.

First, the FocusHandler: this maintains a flag whether the application
is active, i.e., whether any of its Shells is the current active shell.
This needs the workbench to have been created, and the FocusHandler has
acquired over time some completely unrelated code that just also
happened to need a workbench. To ensure that it initialized itself only
once the workbench was ready, the FocusHandler scheduled a job that
re-scheduled periodically until the workbench was up.

Split this: move the unrelated code (setting icons for EGit core jobs)
to the newly introduced StartEventListener. Move maintaining the
"application active" state to ApplicationActiveListener, a dedicated
OSGi component triggered on the APP_STARTUP_COMPLETE event. At that
point, we are guaranteed that the workbench is ready, so no job is
needed. The ApplicationActiveListener fires a custom event
APPLICATION_ACTIVE whenever the active state changes; the data of the
event is a boolean indicating the new state (true if active, false
otherwise). Make sure an initial event is fired. The event is posted
asynchronously.

Move all the repository scanning and project refreshing code to a new
component ExternalRepositoryScanner, which in turn listens on that new
APPLICATION_ACTIVE event. The framework will create this component
automatically when the first such event is fired; component creation
will be off the UI thread since the event is posted asynchronously.
Schedule a run whenever the state changes from "inactive" to "active"
to resume scanning.

Unregister the DebugOptionsListener on Activator.stop().

Change-Id: I296d5801d62a6f2c0c8da4166aa36270c1b2ad1b
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>