Update patch set 3
Patch Set 3:

Thanks for the info, Alex. That's helpful. About the bundle ID hard-coded: yes, I can get the bundle ID at run-time. What I had in mind was this: if I wanted to make the DebugOptionsListener a separate declarative service, I'd end up with something like

  @Component(properties = DebugOptions.LISTENER_SYMBOLICNAME + '=' + "org.eclipse.egit.ui")

and there (and in the XML) I think I would need to hard-code it. So I think leaving that debug registration in the Activator might actually be better.

Patch-set: 3