Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Thanks for the info, Alex. That's helpful. About the bundle ID hard-coded: yes, I can get the bundle ID at run-time. What I had in mind was this: if I wanted to make the DebugOptionsListener a separate declarative service, I'd end up with something like
> 
>   @Component(properties = DebugOptions.LISTENER_SYMBOLICNAME + '=' + "org.eclipse.egit.ui")
> 
> and there (and in the XML) I think I would need to hard-code it. So I think leaving that debug registration in the Activator might actually be better.

That's true, but then if you're doing a refactoring of the bundle name in future (highly unlikely!) then you'll do a refactor which includes hard-coded lines in this project anyway, which will pick it up in the component itself. So hard-coding it is actually the better way.

https://github.com/eclipse/epp.mpc/blob/master/org.eclipse.epp.mpc.ui/OSGI-INF/services/org.eclipse.epp.mpc.ui.debug.options.xml#L3-L5

The ultimate goal would be to try and get rid of the Activator completely :)

Patch-set: 3
Reviewer: Gerrit User 3 <3@97ee7c02-f12f-4043-b43e-dea463d88b31>