Try to stabilize core tests
Looks like the asynchronous file timestamp resolution determination is
sometimes still trying to determine the resolution when a test finishes
and wants to delete the git repository. Avoid this by forcing this to
run synchronously.

Add a lot of debug output when deleting the git directory at the end of
a test fails.

Change-Id: Id79b423ec85eaa0d7d3820ddc79d8d9e24949d50
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>