Update patch set 2
Patch Set 2:

(1 comment)

Sounds good re: future PLUGIN_ID change.

One thing to note: this change will mean EGit is activated earlier than it was before. As soon as the debug plugin loads, it will call this to set up the debug options, which in turn will cause the Activator to be loaded. If we can get the activator's busy work out into separate DS components then this should result in a faster speed-up, but if it's being activated too early you could consider putting in the same block that you have for the workbench ready in the prior component that was migrated.

I mention in case the progress bar of starting Eclipse seems to blip a bit earlier than before, when most of the delay is usually towards the end of the startup.

The good news is that if we manage to evict the Activator code completely with this, then the delay in loading classes won't block the startup at all :)

Patch-set: 2
Reviewer: Gerrit User 3 <3@97ee7c02-f12f-4043-b43e-dea463d88b31>