DiffEditorOutline: simplify presentation toggle
Instead of switching the icon of the toolbar button use a single
image and use IAction.AS_CHECK_BOX. This gives a much more standard
UI. Changing a button's image should be reserved for cases where the
button can have more than two states, as the presentation button in
the staging view or the resource filter button in the history view
toolbar. (Both also have a drop-down menu.)

Use the preference to set the initial state of the button; and update
the preference when the button is toggled. This will make the button
respect the last setting when created (i.e., when a new outline is
opened), but from then on, the states in each outline are independent.

Disable the toolbar buttons if the outline is empty.

Bug: 571032
Change-Id: I487cb24c690896b8e63bd381774071d6df7dea8c
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>