Move more work from Activator to StartEventListener
Both registering the template variable resolver and the configuration
check use jobs that run only once the workbench is up. Simplify this
by moving them to the StartEventListener. By doing so we don't need
a job for the template variable resolvers at all, and the one for the
configuration check becomes much simpler since it doesn't need to
re-schedule if the workbench isn't running yet: it always is.

The configuration check is fairly expensive and involves file system
accesses. Therefore still run it in a job. It doesn't seem to be
necessary to run it on the UI thread, though.

Change-Id: I47a051923df2a2cfd4464f2b0d6984dd0c951d6a
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>