[osgi] Disperse the EGit core Activator into OSGi dynamic services
Use OSGi DS to perform most of the setup formerly done in the Activator.
Because of interdependencies it's not possible to do this step-by-step.

The Activator did a number of unrelated things:

* setting up the internal EGit caches (RepositoryCache, IndexDiffCache,
  RepositoryUtil).
* setting up a number of IResourceChangeListeners for auto-sharing
  projects, auto-ignoring derived resources, and similar things.
* setting up HTTP and SSH, and proxy support.
* handling debug tracing options.
* migrating preferences.
* tracking merge strategies contributed via an extension point.

Each of these have now become an OSGi DS component:

* RepositoryInitializer: EGit caches
* WorkspaceConnector: IResourceChangeListeners
* TransportConfigurator: HTTP, SSH, and proxies
* DebugOptionsHandler: debug tracing options
* PreferencesMigrator: preferences migrations
* MergeStrategyTracker: merge strategy tracking

The activator is no longer the central hub from which other classes
can obtain the RepositoryCache, IndexDiffCache, or RepositoryUtil. This
change was necessary because otherwise an early component activation
causes the Activator to run, which then might cause trying to activate
org.eclipse.core.resources before the instance location is set, which
breaks Eclipse start-up.

These central singletons are newly managed in these three classes
directly, and are accessible via getInstance() static methods.

The RepositoryInitializer publishes the RepositoryCache as a service,
which is consumed by the WorkspaceConnector OSGi component. This
ensures correct initialization order.

The changes outside the Activator and the six new OSGi components are
all simple replacements for using the new instance getters instead of
the Activator.

Bug: 560412
Change-Id: Id5c7440213ae25a573e84720db7dfcc83a8f5d0a
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>