Update patch set 2
Patch Set 2: Code-Review+1

(9 comments)

OK I've taken a walk through the changes and think I understand most of them; some commentary attached.

A few higher-level observations:

* The IWorkspace is created quite early on in the startup of Eclipse's splash screen (probably in the first 10% or so). Moving the dependency on the IWorkspace in the OSGi component will pull the setup quite a lot earlier, so instead of taking time at the ~90% mark you'll find that the ~10% mark it's a bit slower. It will also (paradoxically) make the startup of the resources plugin longer, because when that publishes the IWorkspace, it will trigger components who were waiting on IWorkspace to run.

The above can be remediated somewhat if the 'heavy hitting' OSGi components use the same trick as the StartupInitializer component that was listening for the e4 event and/or Platform.isRunning() checks.

* There's a few lazy-locking type steps (if instance == null, lock() if instance == still null) for static fields. It's almost certainly the case that anyone referring to this class is going to immediately call ThingCache.getInstance() anyway, so actually there's not much benefit in doing that versus just having the instance created as a static final field; and that way, you know it's not null, so you can avoid a bunch of `if null` checks. You might not want/need to do that change in this one as it can be done afterwards as a subsequent change.

* The thing I'm mostly concerned on is the RepositoryUtil.getInstance(), which blocks until such time as an instance is available. This is used in a few places (noted elsewhere) which is potentially gong to run on the main UI thread, including dialogs being shown, classes being loaded etc.

I understand that the issue with RepositoryUtil is that it won't exist until the location has been set, but perhaps another way of doing it is to have a RepositoryUtil that can set the location as a mutable field, and then have the implementations of (e.g.) 'contains' return null or 'getName' to be an empty string.

You'll have more knowledge of where/how/why RepositoryUtil is used, and might be able to think about other approaches that would work in this case, but that is the most contentious part of the change that I can think of.

Patch-set: 2
Reviewer: Gerrit User 3 <3@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=+1