Update patch set 4
Patch Set 4:

> I have the impression that loading these strategies may cause *a lot* of bundle activation as it even instantiates the classes for the contributed strategies.

Unfortunately this extension point is badly designed. It has a "name" attribute, but it's optional. If it were mandatory, class loading would not be necessary as far as I see. But at least EMF compare, which introduced this extension point, doesn't set that optional "name" attribute.[1] And changing it would be a breaking API change :-( If it was only EMF compare that used this, we might be able to coordinate changing this extension point. But who knows which other bundles may have made use of this since it was introduced in 2015?

@Laurent (both of you :-)): any suggestions how we might improve this?

[1] https://git.eclipse.org/r/plugins/gitiles/emfcompare/org.eclipse.emf.compare/+/refs/heads/master/plugins/org.eclipse.emf.compare.egit/plugin.xml

Patch-set: 4
CC: Gerrit User 6332 <6332@97ee7c02-f12f-4043-b43e-dea463d88b31>
CC: Gerrit User 934 <934@97ee7c02-f12f-4043-b43e-dea463d88b31>