Update patch set 4
Patch Set 4:

I like the idea of moving the requests to Jobs, especially with the cancel for when it shuts down. Getting long-running jobs out of the Activator is going to help startup speed, in that a number of them can start in parallel.

If they're starting too early, we can insert some delays into the OSGi components by listening to the e4 ready message topic; this will have the effect of deferring them until needed. I think this will be appropriate for things that are dependant upon registering the resource change listeners.

Patch-set: 4
Reviewer: Gerrit User 3 <3@97ee7c02-f12f-4043-b43e-dea463d88b31>