Update patch set 7
Patch Set 7:

I'd like to know all the same what the debug trace shows when you use the command-line prompt. If it's something we can recognize with reasonable certainty, it might be worth checking for this to produce an exception with a message that can give enough hints to the end user so that s/he can fix the problem him- or herself.

Patch-set: 7