Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> I'd like to know all the same what the debug trace shows when you use the command-line prompt. If it's something we can recognize with reasonable certainty, it might be worth checking for this to produce an exception with a message that can give enough hints to the end user so that s/he can fix the problem him- or herself.

If I set GPG_TTY and then run gpg on the command-line, it reports

  PINENTRY_LAUNCHED 22712 mac 1.1.0 /dev/ttys000 xterm-256color /private/tmp/com.apple.launchd.jobL5VbdW6/org.macosforge.xquartz:0

and still shows the dialog. Fields 5 and 6 seem to relate to this tty setting. Field 7 is $DISPLAY. Interestingly the --no-tty option that is passed does not suppress this tty info, and pinentry-mac shows the dialog even if unset $DISPLAY!

Patch-set: 7