Update patch set 7
Patch Set 7:

>  [GNUPG:] PINENTRY_LAUNCHED 3175 curses 1.1.1 - xterm-256color /private/tmp/com.apple.launchd.EFYmOHTFKI/org.macosforge.xquartz:0 - 503/20 0
>  gpg: signing failed: Inappropriate ioctl for device
> [GNUPG:] FAILURE sign 83918950
>  gpg: signing failed: Inappropriate ioctl for device
> 
> I sent the complete trace via email

At least we could detect that gpg-agent use pinentry-curses or pinentry-tty. But that still doesn't really help if a user has a gpg-agent that is started in way that it always uses that. Yeah, unsetting GPG_TTY won't help of course if gpg-agent is running already. Confused myself momentarily there.

The only way (I know of) to make gpg-agent choose per request is a custom script as pinentry program that chooses the real pinentry program based on the value of the PINENTRY_USER_DATA environment variable, which is passed through from gpg to gpg-agent to the pinentry. Such a setup is way too complicated for most users.

And if pinentry-curses fails already with "inappropriate ioctl", even attempting to parse stdout and feed a passphrase via stdin won't work.

Patch-set: 7