GPG signing: preference to choose between BC and external GPG
Like we do for HTTP, and like we did for SSH: give the user a
preference to choose which implementation is to be used for signing
commits or tags. Default (for now) is the Bouncy Castle library.

Add the preference to the "Committing" preference page.

JGit-Dependency: I6201f98d4b1704a053159967b8adacd98e368522
Change-Id: Iedf0209635e903b2919d6ec4d201aaa389391d1e
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>