GPG signing: detect GPG trying to do tty input
Some GPG installations may be configured to request user input for
passphrases or smartcard PINs via a terminal. When EGit invokes GPG,
there is no terminal, and the gpg command fails.

Detect this case and throw an exception with a descriptive message.
Report this error to the user in a dedicated ErrorDialog that shows
a "help" button. Set a help context, and provide a dedicated help
page explaining possible causes of and potential remedies for the
problem.

Change-Id: Idd95d1394b6fbb0f443e0efa7ec8cb3940236c1f
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>