Update patch set 5
Patch Set 5:

Refactored this merge strategy business. Now in class MergeStrategies; still with a job as OSGi component to load. If merge strategies are requested before the job has initialized the strategies, empty results are returned. Removed the remnants of this merge strategy business from the Activator; now also in class MergeStrategies.

Still not happy with that extension point for merge strategies, but at least it's completely outside of the Activator now.

The Activator now only does some very basic but very important JGit setup that has to happen very early on, and provides logging utilities.

Perhaps this logging business could also be done differently, but that's for a later change. It would incur changes all over the place.

Patch-set: 5