Update patch set 6
Patch Set 6:

> should this wait for the next release or do you want to get this in for 5.11 ?

IMO this should go in early in the next release. While I'm confident it won't cause any problems, it's simply safer: if there are problems all the same, analyzing them will be easier with only one Activator (UI) changed in one release and the other in the next release. I see no reason to rush this.

It'll also give the EMF Compare team more time to detect possible problems with the asynchronous merge strategy loading, or to consider improving that extension point.

Or do you think that's overly cautious?

Patch-set: 6