Update patch set 6
Patch Set 6:

Sorry for the lack of response from our (EMF Compare) side. Thanks a lot for the heads up, we just haven't had any time to consider the changes here and how to enhance the extension point on our end. I'll try and look into this asap but don't think I'll have any chance to do so for 2021-03.

Patch-set: 6